Move storage binds and slot click handling to shared
